Refactor(client): HASHI-156 RestaurantDetailPage 책임 분리
