+++ apps/client/src/features/restaurantDetail/hooks/useRestaurantDetailReviewUnavailableModal.ts
@@ -1,0 +1,25 @@
+import { useState } from 'react'
+
+export const useRestaurantDetailReviewUnavailableModal = () => {
+  const [isReviewUnavailableModalOpen, setIsReviewUnavailableModalOpen] =
+    useState(false)
+
+  const openReviewUnavailableModal = () => {
+    setIsReviewUnavailableModalOpen(true)
+  }
+
+  const closeReviewUnavailableModal = () => {
+    setIsReviewUnavailableModalOpen(false)
+  }
+
+  const resetReviewUnavailableModal = () => {
+    setIsReviewUnavailableModalOpen(false)
+  }
+
+  return {
+    isReviewUnavailableModalOpen,
+    onCloseReviewUnavailableModal: closeReviewUnavailableModal,
+    onOpenReviewUnavailableModal: openReviewUnavailableModal,
+    resetReviewUnavailableModal,
+  }
+}
